chore(): move utils
